requirements.txt
# QuickDraw Sketch Recognition API
# Compatible with Python 3.10+ on Windows, macOS (Intel & Apple Silicon), and Linux

# Core dependencies
fastapi>=0.115.2
uvicorn[standard]>=0.24.0
pydantic>=2.7.4
python-multipart>=0.0.18

# ML/AI libraries
tensorflow>=2.15.0
numpy>=1.25.0,<2.0  # TensorFlow 2.15 requires numpy < 2.0
scikit-learn>=1.3.2
matplotlib>=3.8.2

# Image processing
Pillow>=10.1.0

# ONNX support (optional, for model export)
tf2onnx>=1.15.1
onnx>=1.15.0
onnxruntime>=1.16.3

# Development and testing
pytest>=7.4.3
httpx>=0.25.2
requests>=2.32.2
